* src/pluginlaunch.c (pluginlaunch_stop): Add soft_stop argument.
* src/pluginlaunch.h: Adjust function parameter.

* src/attack.c (launch_plugin, attack_host, handle_scan_stop_signal):
Adjust function call.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-scanner@26467 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/pluginlaunch.c
@@ -376,18 +376,22 @@ pluginlaunch_enable_parrallel_checks (void)
 
 
 void
-pluginlaunch_stop (void)
+pluginlaunch_stop (int soft_stop)
 {
   int i;
-  read_running_processes ();
 
-  for (i = 0; i < MAX_PROCESSES; i++)
+  if (soft_stop)
     {
-      if (processes[i].pid > 0)
-        kill (processes[i].pid, SIGTERM);
+      read_running_processes ();
+
+      for (i = 0; i < MAX_PROCESSES; i++)
+        {
+          if (processes[i].pid > 0)
+            kill (processes[i].pid, SIGTERM);
+        }
+      usleep (20000);
     }
 
-  usleep (20000);
   for (i = 0; i < MAX_PROCESSES; i++)
     {
       if (processes[i].pid > 0)
